Documentation and formatting update
