+++ pillarbox-core-business/src/main/java/ch/srgssr/pillarbox/core/business/source/DefaultMediaMetaDataProvider.kt
@@ -16,23 +16,23 @@ import ch.srgssr.pillarbox.player.extension.setCredits
  * Filling [MediaMetadata] from [Chapter].
  * [MediaMetadata] provided fields are not replaced.
  */
-val DefaultMediaMetaDataProvider: suspend MediaMetadata.Builder.(MediaMetadata, Chapter, MediaComposition) -> Unit = { metadata,
-        chapter, mediaComposition ->
-    metadata.title ?: setTitle(chapter.title)
-    metadata.subtitle ?: setSubtitle(chapter.lead)
-    metadata.description ?: setDescription(chapter.description)
-    metadata.artworkUri ?: run {
-        val artworkUri = imageScalingService.getScaledImageUrl(
-            imageUrl = chapter.imageUrl
-        ).toUri()
-        setArtworkUri(artworkUri)
+val DefaultMediaMetaDataProvider: suspend MediaMetadata.Builder.(MediaMetadata, Chapter, MediaComposition) -> Unit =
+    { metadata, chapter, mediaComposition ->
+        metadata.title ?: setTitle(chapter.title)
+        metadata.subtitle ?: setSubtitle(chapter.lead)
+        metadata.description ?: setDescription(chapter.description)
+        metadata.artworkUri ?: run {
+            val artworkUri = imageScalingService.getScaledImageUrl(
+                imageUrl = chapter.imageUrl
+            ).toUri()
+            setArtworkUri(artworkUri)
+        }
+        ChapterAdapter.getChapters(mediaComposition)
+            .takeIf { it.isNotEmpty() }
+            ?.let { setChapters(it) }
+        TimeIntervalAdapter.getCredits(chapter.timeIntervalList)
+            .takeIf { it.isNotEmpty() }
+            ?.let { setCredits(it) }
     }
-    ChapterAdapter.getChapters(mediaComposition).takeIf { it.isNotEmpty() }?.let {
-        setChapters(it)
-    }
-    TimeIntervalAdapter.getCredits(chapter.timeIntervalList).takeIf { it.isNotEmpty() }?.let {
-        setCredits(it)
-    }
-}
 
 private val imageScalingService = ImageScalingService()
